Done with Jupitor notebook
Some of the most basic and important Preprocessing techniques are:-

1) Binarization
2) Skew Correction
3) Noise Removal
4) Thinning and Skeletonization

Binarization: In layman’s terms Binarization means converting a coloured image into an image which consists of only black and white pixels (Black pixel value=0 and White pixel value=255). As a basic rule, this can be done by fixing a threshold (normally threshold=127, as it is exactly half of the pixel range 0–255). If the pixel value is greater than the threshold, it is considered as a white pixel, else considered as a black pixel.

But this strategy may not always give us desired results. In the cases where lighting conditions are not uniform in the image, this method fails.

But this strategy may not always give us desired results. In the cases where lighting conditions are not uniform in the image, this method fails.

Imax= Maximum pixel value in the image, Imin= Minimum pixel value in the image, E = Constant value Source:

2. Skew Correction: While scanning a document, it might be slightly skewed (image aligned at a certain angle with horizontal) sometimes. While extracting the information from the scanned image, detecting & correcting the skew is crucial.
Several techniques are used for skew correction.

→ Projection profile method
→ Hough transformation method
→ Topline method
→ Scanline method

However, the projection profile method is the simplest, easiest and most widely used way to determine skew in documents.

In this method, First, we’ll take the binary image, then

project it horizontally (taking the sum of pixels along rows of the image matrix) to get a histogram of pixels along the height of the image i.e count of foreground pixels for every row.
Now the image is rotated at various angles (at a small interval of angles called Delta) and the difference between the peaks will be calculated (Variance can also be used as one of the metrics). The angle at which the maximum difference between peaks (or Variance) is found, that corresponding angle will be the Skew angle for the image.
After finding the Skew angle, we can correct the skewness by rotating the image through an angle equal to the skew angle in the opposite direction of skew.


3. Noise Removal: The main objective of the Noise removal stage is to smoothen the image by removing small dots/patches which have high intensity than the rest of the image. Noise removal can be performed for both Coloured and Binary images.
One way of performing Noise removal by using OpenCV fastNlMeansDenoisingColored function.

More about Noise removal & Image smoothening techniques can be found in this wonderful article

4. Thinning and Skeletonization: This is an optional preprocessing task which depends on the context in which the OCR is being used.
→ If we are using the OCR system for the printed text, No need of performing this task because the printed text always has a uniform stroke width.
→ If we are using the OCR system for handwritten text, this task has to be performed since different writers have a different style of writing and hence different stroke width. So to make the width of strokes uniform, we have to perform Thinning and Skeletonization.